Import United Learning vacancies
Adds the ability to import vacancies from external data sources such as
XML feeds or APIs, and adds an initial feed import for a trial launch.

This will probably be extended and refactored extensively in the future
once we have additional data sources that we integrate with and can
abstract/extract common code.

- Add a concept of external sources to import data from
- Add an initial data source for United Learning based on an XML feed
- Add some display tweaks for imported vacancies
- Add a job for running the import and add it to the schedule